moved lcd backlight's stuff to separate class, fixed compiler warnings in HAL

+++ Firmware/Drivers/STM32F4xx_HAL_Driver/Inc/stm32f4xx_hal_def.h
@@ -81,7 +81,7 @@ typedef enum
                               (__DMA_HANDLE__).Parent = (__HANDLE__);             \
                           } while(0)
 
-#define UNUSED(x) ((void)(x))
+#define UNUSED(x) ((void)((uint32_t)(x)))
 
 /** @brief Reset the Handle's State field.
   * @param __HANDLE__ specifies the Peripheral Handle.
